requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Login
Flask-JWT-Extended
paho-mqtt
plotly
Werkzeug
SQLAlchemy